Add files via upload
Here's the text file for the answers. Let me know what I need to change. I was going to put everything in a js file - guess I still could have - but wanted to make sure we showed we can parse and do this with AJAX. If I didn't do what was expected, please let me know.