Here's the json file.
All the responses are in this one file, but I do have 2 different objects in here. I did it just in case we need to use a type (as in if it's neutral, negative, or positive), but we don't have to use it. Also, I'm uploading this file like this because I'm still having issues with pushing in netbeans. Anyway, at least this is here. Working on AJAX.