Still not working
So, this is what I've written but can't get to work. I've checked the code validator and even the json file formatting, but they all give me a green light. The error I get when I debug says that my loadJson function isn't defined, but I can't figure out how when everything's in there. Can someone please take a look with a fresh pair of eyes to see what I may be missing please? I've been combing through it and have found errors, but this is the cleanest version. In the meantime, I'll continue to look through it. Maybe, I mistakenly named something. I am so sorry for how long this has taken.